+++ src/main/java/techcourse/myblog/article/domain/Article.java
@@ -1,0 +1,33 @@
+package techcourse.myblog.article.domain;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Article {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    private String title;
+
+    private String coverUrl;
+
+    private String contents;
+
+    @Builder
+    private Article(long id, String title, String coverUrl, String contents) {
+        this.id = id;
+        this.title = title;
+        this.coverUrl = coverUrl;
+        this.contents = contents;
+    }
+}
