+++ src/main/java/techcourse/myblog/article/domain/ArticleRepository.java
@@ -1,0 +1,6 @@
+package techcourse.myblog.article.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface ArticleRepository extends CrudRepository<Article, Long> {
+}
