+++ src/main/java/techcourse/myblog/article/exception/NotFoundArticleException.java
@@ -1,0 +1,7 @@
+package techcourse.myblog.article.exception;
+
+public class NotFoundArticleException extends IllegalArgumentException {
+    public NotFoundArticleException() {
+        super("찾을 수 없는 게시물입니다.");
+    }
+}
