--- src/main/java/techcourse/myblog/domain/ArticleRepository.java
@@ -1,6 +1,0 @@
-package techcourse.myblog.domain;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface ArticleRepository extends CrudRepository<Article, Long> {
-}
