--- src/main/java/techcourse/myblog/domain/UserRepository.java
@@ -1,9 +1,0 @@
-package techcourse.myblog.domain;
-
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends CrudRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
