--- src/main/java/techcourse/myblog/exception/DuplicatedUserException.java
@@ -1,7 +1,0 @@
-package techcourse.myblog.exception;
-
-public class DuplicatedUserException extends IllegalArgumentException {
-    public DuplicatedUserException() {
-        super("중복된 e-mail입니다.");
-    }
-}
