--- src/main/java/techcourse/myblog/exception/InvalidEditFormException.java
@@ -1,7 +1,0 @@
-package techcourse.myblog.exception;
-
-public class InvalidEditFormException extends RuntimeException {
-    public InvalidEditFormException(String message) {
-        super(message);
-    }
-}
