--- src/main/java/techcourse/myblog/exception/InvalidSignUpFormException.java
@@ -1,7 +1,0 @@
-package techcourse.myblog.exception;
-
-public class InvalidSignUpFormException extends RuntimeException {
-    public InvalidSignUpFormException(String message) {
-        super(message);
-    }
-}
