--- src/main/java/techcourse/myblog/exception/NotFoundUserException.java
@@ -1,7 +1,0 @@
-package techcourse.myblog.exception;
-
-public class NotFoundUserException extends RuntimeException {
-    public NotFoundUserException() {
-        super("아이디가 존재하지 않습니다.");
-    }
-}
