+++ src/main/java/techcourse/myblog/user/domain/UserRepository.java
@@ -1,0 +1,9 @@
+package techcourse.myblog.user.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends CrudRepository<techcourse.myblog.user.domain.User, Long> {
+    Optional<techcourse.myblog.user.domain.User> findByEmail(String email);
+}
