+++ src/main/java/techcourse/myblog/user/exception/InvalidEditFormException.java
@@ -1,0 +1,7 @@
+package techcourse.myblog.user.exception;
+
+public class InvalidEditFormException extends RuntimeException {
+    public InvalidEditFormException(String message) {
+        super(message);
+    }
+}
