+++ src/main/java/techcourse/myblog/user/exception/InvalidSignUpFormException.java
@@ -1,0 +1,7 @@
+package techcourse.myblog.user.exception;
+
+public class InvalidSignUpFormException extends RuntimeException {
+    public InvalidSignUpFormException(String message) {
+        super(message);
+    }
+}
