+++ src/main/java/techcourse/myblog/user/exception/NotFoundUserException.java
@@ -1,0 +1,7 @@
+package techcourse.myblog.user.exception;
+
+public class NotFoundUserException extends RuntimeException {
+    public NotFoundUserException() {
+        super("아이디가 존재하지 않습니다.");
+    }
+}
