--- src/test/java/techcourse/myblog/web/LoginControllerTests.java
@@ -1,74 +1,0 @@
-package techcourse.myblog.web;
-
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.web.reactive.server.WebTestClient;
-import org.springframework.web.reactive.function.BodyInserters;
-
-@ExtendWith(SpringExtension.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class LoginControllerTests {
-    @Autowired
-    private WebTestClient webTestClient;
-
-    @BeforeEach
-    void setUp() {
-        webTestClient.post().uri("/users")
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-                .body(BodyInserters
-                        .fromFormData("email", "email@gmail.com")
-                        .with("password", "password1234!")
-                        .with("name", "name"))
-                .exchange();
-    }
-
-    @Test
-    void 로그인_페이지_이동_테스트() {
-        webTestClient.get().uri("/login")
-                .exchange()
-                .expectStatus().isOk();
-    }
-
-    @Test
-    void 로그인_요청_성공_테스트() {
-        webTestClient.post().uri("/login")
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-                .body(BodyInserters
-                        .fromFormData("email", "email@gmail.com")
-                        .with("password", "password1234!"))
-                .exchange()
-                .expectStatus().isFound();
-    }
-
-    @Test
-    void 로그인_요청_존재하지_않는_이메일_테스트() {
-        String noneEmail = "none@gamil.com";
-
-        webTestClient.post().uri("/login")
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-                .body(BodyInserters
-                        .fromFormData("email", noneEmail)
-                        .with("password", "password"))
-                .exchange()
-                .expectStatus().isBadRequest();
-    }
-
-    @Test
-    void 로그인_요청_패스워드_불일치_테스트() {
-        String wrongPassword = "wrongPassword";
-
-        webTestClient.post().uri("/login")
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-                .body(BodyInserters
-                        .fromFormData("email", "email@gmail.com")
-                        .with("password", wrongPassword))
-                .exchange()
-                .expectStatus().isBadRequest();
-    }
-}
